Add support Vue 3 + update .gitignore
