+++ .gitignore
@@ -1,3 +1,25 @@
-/node_modules
+# Dependency directories
+node_modules
+coverage/
+package-lock.json
+*.vim
 
-npm-debug.log
+# Yarn Integrity file
+.yarn-integrity
+
+# Logs
+/logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# IDE / Editor
+.idea
+.vscode
+
+# macOS
+.DS_Store
+
+# Optional eslint cache
+.eslintcache
